+++ librds/datatypes.py
@@ -19,10 +19,9 @@ class GroupIdentifier:
     group_number: int
     group_version: bool
 
-
 @dataclass
 class Details:
-    details=True
+    details=""
 
 @dataclass
 class PSDetails(Details):
@@ -35,6 +34,7 @@ class PSDetails(Details):
     ta: bool
     text: str
     af: AlternativeFrequencyEntryDecoded
+    details="PS"
 
 @dataclass
 class FastSwitchingInformation(Details):
@@ -45,32 +45,38 @@ class FastSwitchingInformation(Details):
     di_dpty:bool
     ms: bool
     ta: bool
+    details="FSI"
 
 @dataclass
 class RTDetails(Details):
     segment: int
     ab: bool
     text: str
+    details="RT"
 
 @dataclass
 class LongPSDetails(Details):
     segment: int
     text: str
+    details="LPS"
 
 @dataclass
 class PTYNDetails(Details):
     segment: int
     ab: bool
     text: str
+    details="PTYN"
 
 @dataclass
 class PINSLCetails(Details):
     data:int
     is_lic:bool
+    is_broadcaster_data:bool
     variant_code: int
     pin_day: int
     pin_hour: int
     pin_minute: int
+    details="PINSLC"
 
 @dataclass
 class CTDetails(Details):
@@ -79,31 +85,37 @@ class CTDetails(Details):
     minute: int
     time_sense: bool
     local_offset: int
+    details="CT"
 
 @dataclass
 class TDCDetails(Details):
     channel: int
     data: list[int]
+    details="TDC"
 
 @dataclass
 class InHouseDetails(Details):
     data: list[int]
+    details="IH"
 
 @dataclass
 class ODADetails(Details):
     data: list[int]
+    details="ODA"
 
 @dataclass
 class ODAAidDetails(Details):
     oda_group: GroupIdentifier
     aid:int
     scb:int
+    details="ODAAID"
 
 @dataclass
 class EONBDetails(Details):
     pi: int
     tp: bool
     ta: bool
+    details="EONB"
 
 @dataclass
 class EONADetails(Details):
@@ -115,12 +127,13 @@ class EONADetails(Details):
     pty: int
     on_af: int
     variant_code:int
+    details="EONA"
 
 @dataclass
 class DecodedGroup:
-    raw_group:Group
+    raw_group: Group
     pi: int
     tp: bool
     pty: int
     group: GroupIdentifier
-    details:Details
+    details: Details